feat(ai): split copilot and vscode
